refactor: Meld flavor and and circuit builder modules (#5406)
The proof_system module has become a pile of things, and the name is
useless. The PR resolves that and moves us toward a model where flavor
and circuit builders live side-by-side since this is how we implement
check_circuit.

Split proof_system module. 
 - Ultra & GoblinUltra go into into stdlib_circuit_builders
- VM trace builders go into single folder alongside other stuff (no
stdlib dependency, so little benefit to building separately)
- Library stuff goes into plonk_honk_shared (library stuff); will
refactor when Plonk moves out.
 - Polynomials module gets the polynomial store.

We also simplify a little bit of the linking structure.
New module structure:
![barretenberg_modules](https://github.com/AztecProtocol/aztec-packages/assets/26756572/a773b29d-370a-434e-baa7-7d9780942a28)