fix: Correct circuit size estimation for UltraHonk (#6164)
Resolves https://github.com/AztecProtocol/aztec-packages/issues/6161

Look into the issue for more context on the problem.

I moved `add_gates_to_ensure_all_polys_are_non_zero` to be part of
`finalize_circuit` for both the goblin and ultra builder. Then upon
creating an AcirFormat circuit we now call finalize_circuit. Previously
we were having to attach an additional buffer as a hack as we were using
`get_total_circuit_size` which correctly estimated a finalized circuit
size for plonk, but not for Honk. We will now have these additional
gates for UltraPlonk as well, but this is fine as they are not that many
gates (~12 gates) and we are removing plonk anyway.