feat: Use actual tx fee in gas token when charging fee (#6166)
Uses the actual `transaction_fee` injected during teardown for `pay_fee`
in the `GasToken`.

This PR also refactors the e2e_fees and e2e_dapp_subscription tests to
use snapshots, and changes dapp_subscription so each test can be run
independently. It also modifies the `AppSubscription` contract so the
max fee it covers is configurable during initialization, instead of
hardcoded to 42. It also lowers the severity of some random logs to make
e2e output more readable (mostly around prover). Last, it includes
bugfix for the snapshot manager when it runs in "disabled mode", where
it was recreating the context multiple times, and refactors it by
splitting it into different classes depending on the mode it runs.