fix: increase N_max in Zeromorph (#6415)
To avoid edge case in `batch_mul` of vanilla Ultra we temporarily
switched to naive muls in the folding recursive verifier circuit
arithmetised with Ultra which causes an increase by a factor of 2. Hence
we need to increase the maximum size of a circuit that can be proven
with Zeromorph.