fix: Do not use kernel constants for computing tx fee (#6635)
We were using the `CombinedConstantData` from the kernel public inputs
for getting the block fees and computing the txFee. However, the
`globals.gasFees` in the `CombinedConstantData` that come from a
`PrivateKernelTail` are empty, so we were always computing a zero fee
for txs without a public component.

This PR changes it so we always have to inject the block gas fees to
compute the actual tx fee.

---

This hints at another problem we may have: we have fields in our
circuits structs that are sometimes safe to use, and sometimes not.
Should we rethink our structs so we don't accidentally run into this
situation again? Another example that comes to mind is the
`transaction_fee` field which is only set during teardown.