feat: Changing finite field arithmetic in wasm to 29 bits for multiplications (second try, disabled AVM build in wasm) (#6027)
Now when we compile for wasm we use 9 29-bit limbs to perform all
multiplications. In wasmtime this results in -10% for multiplication
benchmark and -35% in squaring. This makes a -7% impact on wasm
client_ivc.
Other changes:
1. Most of the changes are just additions to tests to handle the new
montgomery form, because we have a lot of hardcoded internal
representations. If we get a new one, we should rewrite those to use
non-montgomery form. Also we need to recompute all parameters.
2. Added squarings to univariate and used in some relations
3. Added a script to run wasmer instead of wasmtime on the benchmark
(but you need to change arguments slightly to add "--")
4. For some reason wasm build has become really slow (especially on
relations and protogalaxy).
5. Added docs

![image](https://github.com/AztecProtocol/aztec-packages/assets/4798775/7e28bbab-6735-4309-aa82-cf81dd813d18)

Before:

![image](https://github.com/AztecProtocol/aztec-packages/assets/4798775/15f48eb2-9343-4d8c-a444-a37fd5b6f3c8)

---------

Co-authored-by: ludamad <adam.domurad@gmail.com>