feat: allow get_notes to return zero notes (#7621)
`get_notes` used to fail if no notes were found ever since
https://github.com/AztecProtocol/aztec-packages/issues/4988 was merged.
The argument there was that since no constraints are applied, and note
non-existence cannot be proved, `get_notes` returning no notes is always
a mistake.

However, what that mistake means depends on the caller. Since we don't
want for the caller to have to pass an error message to `get_notes`, I
think it's better to return no notes and leave it up to the caller to
decide what to do. `BoundedVec` is a safe container: `get` will result
in errors if used with indices out of bounds.

I changed the callsites where we did `get_unchecked(0)` to `get(0)` -
this means that the errors now change from 'got 0 notes' to 'out of
bounds access'. Neither error message is great, so I didn't bother
improving this. Maybe we could have a `get` variant that received an
error message (like `Option::expect`), but we don't need to worry about
this for now.

Closes https://github.com/AztecProtocol/aztec-packages/issues/7059
(indirectly due to how we need to change certain tests to address this)