fix(avm): correctly build spike vm (#7726)
After this PR you can now generate an alternative (spike) vm and it will
build correctly. Most of the changes are to account for empty entity
subsets.

I had to add some typing information to `flavor_macros.hpp` to work when
the flavor is empty. This is annoying but shouldn't have any runtime
implications.