fix: new commit_sparse bug and new tests (#8649)
This bug in commit_sparse was due to an assumption that the size of the
polynomial was a power of 2. The structured polynomials work recently
made this assumption incorrect as we can set start and end indices for a
polynomial. One note is that the bug only happens if the size is large
enough (bigger than the number of cpus from `get_num_cpus_pow2()`)
because we otherwise would use 1 thread (which divides any number).

The 3 new tests test commit_sparse on a polynomial that has a size !=
virtual_size, a start_idx != 0, and a larger test case where both are
true. This larger test case fails without the fix.