chore!: change ec_add to unsafe implementation (but much better perf) (#8374)
Use `unconditional_add` for ec_add ACIR opcode.
This improves a lot the performance of the opcode, but also makes it
unsafe.
~~I keep the PR as draft until aztec packages are sync with Noir PR
https://github.com/noir-lang/noir/pull/5858 which adds the checks in the
stdlib function.~~
The unsafe version can then be used when we know the inputs are valid
(for instance if they come from a previous add).
n.b.: the real performance boost will happen when we will be able to use
the unsafe version.

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>