feat: Stronger typing in L1 contracts (#8841)
Bumps the solidity version and makes use of the user defined types such
that we have `Timestamp`, `Slot` and `Epoch` to make it less likely
mixing up these `uint256` that have different "units".

Spurred from a timestamp issues, as a slotnumber was passed along and we
should be able to make it scream earlier with this.

Note that as the types are mostly syntactic sugar and won't make their
way all the way down into the abis the errors and selectors should stay
the same.