fix: flaky e2e sample dapp and quick start (#8768)
This fixes the flakiness of `up_quick_start` and `sample_dapp`. It also
updates the test using the cli to the correct version.

One notable source of flakiness that I chased down was a weird`Assertion
failed: membership check failed`. This was showing up pretty reliably in
`sample_dapp`. I solved this by separating the tests, and confirmed by
running it in band (thanks @Thunkar) because it seems like crosstalk
between both tests hitting the same pxe is a trigger of this. I explored
but did not find the actual cause on why the crosstalk would induce this
error, as I had spent already too much time but I have made an issue and
have linked it in the code.