feat(avm): trace contract class and contract instance (#8840)
This PR is centred around tracing and passing contract class & instance
during simulator execution and passing it to circuit. We store each
contract class & instance whenever the `simulator` calls `getBytecode`.

This changes the input interface to the bb binary - we no longer take in
a specific bytecode to execute. Instead we get a vector of
`{contract_class, contract_instance, bytecode}` which define all the
(deduplicated) contract bytecode that will be executed during this
"one-enqueued call" (actual implementation of 1-enqueued call tbd).

This doesnt do any derivation of id or address yet