refactor: add sender to encode and encrypt (#9562)
This PR is purely plumbing, as we add a sender to encode_and_encrypt,
changing the API because in the future we will require the sender as a
param to request the shared secret from the PXE that will be used to
compute the tag.

We are using a placeholder of `msg_sender` and trying to infer from
context who the sender should be whenever possible.

For the fee / partial note stuff, it may be good to have a better
understanding who the sender is.

Finally, some patterns arise, namely that we should probably add self as
an implicit member of the address book :D.

To note: this pollutes the API even more, and is beyond the standard of
sane. as discussed on Slack, we should abstract this pollution wherever
possible.