chore: fixing test contract fixture (#9909)
This just fixes a utility function that seems to be a typo, as it is has
the same content as the function above it, and uses the same artifact.
This also bumps the timeout of `pxe.test.ts` because we were seeing some
issues in timeouts with adding contract artifact.