chore: build acir test programs in parallel to e2e build step (#9988)
This PR shaves about a minute off of the `acir-test`s jobs as we can
build all of the ACIR programs in parallel with the main `build` step.
I've decided to add an extra dependency on `build` so that we're only
changing when the ACIR is generated but still running the tests at the
same time.

The `noir+build-acir-tests` target only depends on the `+nargo` target
and nothing outside.