refactor: remove PublicExecutor (#10028)
- Pulls what the PublicExecutor did into an internal function of the
PublicTxSimulator
- Modifies the TXE to craft a carrier TX for public calls and just call
`publicTxSimulator.simulate(tx)` so that we only need to maintain that
interface for external use of the simulator
- Adds `gasLeft` to `AvmContractCallResult`, removes gasLeft as an extra
param to functions that accept call result.
- Introduces `AvmFinalizedCallResult` which is just like
`AvmContractCallResult` but is friendlier for use externally to the AVM.
Uses standard Gas type, and uses `SimulationError` type for
`revertReason`.
- Note that we might want to only have one gas type eventually, but even
just having this dedicated result type with revertReason of type
SimulationError led to some cleanup
- Removes bytecode size metering that was done in PublicExector (it
wasn't useful anymore)
- Removes unused/stale SimulationProvider paramters