test: Fast epoch building test (#10045)
Adds a fast (as in 1 minute) e2e test for building epochs. Checks that
the epoch is properly reorg'd, and that the prover proof cannot land
outside the epoch even if not blocks have been submitted.

- Introduces a tx delayer that lets us pick in which L1 block or
timestamp a transaction sent by the L1 publisher will land.
- Includes the fix from #10084 as a cherry-picked commit.
- Allows properly setting slot and epoch duration in e2e setup.
- Adds a `LOG_ELAPSED_TIME` env var to log the total elapsed times in
debug logs.
- Adds a `rollup` wrapper for the Rollup contract, so we don't need to
deal with viem explicitly.
- Moves `startAnvil` to the `aztec/ethereum` package

Fixes #9809