chore: update external joiner script for new networks (#10810)
The script now takes an explicit reference to a network.

Networks have hardcoded configurations to loadbalancers that sit in
front of boot-nodes and the ethereum host.

For now, only `troll-turtle` is supported. Users can override to join
custom networks by setting.

Note: this means that if the `troll-turtle` network goes down, these
test will fail.

I'm fine with that for now, but I filed
https://github.com/AztecProtocol/aztec-packages/issues/10809