fix: Race condition when cleaning epoch proof quotes (#10795)
The test `In-Memory P2P Client › deletes expired proof quotes` was
sometimes failing with `epochProofQuotePool.deleteQuotesToEpoch` not
being called.

This could happen on a race condition where the p2p client reports to
have synched to a given block number, but the quotes are deleted *after*
the synched block number is set.

This PR ensures setting the last block number is the last operation
done.