fix: boxes playwright + web bundles (#11615)
Don't really get why the correct version is not being installed

@ludamad pulling `aztecprotocol/build:2.0` locally reveals the
playwright version is wrong, which explains all the failures. Maybe I
messed up badly by not bumping the image version? Can we fix it without
pushing for 3.0? I added a stopgap measure in the boxes docker-compose,
but it takes way too long as it is, so would love to fix it for good.