fix: Add missing return in main (#11786)
https://github.com/AztecProtocol/aztec-packages/pull/10217 introduced a
bug where the verifier result would not be sent from bb main as a
signal. This triggers a mysterious error in the ivc-integration tests
suite, native only, but does not break any e2e tests. More investigation
is necessary, and ultimately this should be resolved in the ACIR tests,
but it feels prudent now to fix the bug and reinstate the tests later.