feat: batch writes to the proving broker database (#11900)
This PR adds `BatchQueue` which groups items together by a custom 'group
key' and then processes them as a batch in a loop.

This queue is then used in the prover broker to group together writes
(new proving jobs and proving results) in order to achieve greater write
throughput (batches go in as a single write transaction).

This PR supersedes #11784