refactor: remove `addNullifiedNote` from pxe (#11822)
This PR removes `addNullifiedNote` from the PXE as it is unused.

It was also suggested to rework `getNoteHashAndOptionallyANullifier`
into `getNoteHashAndNullifier`, in TS as now the second case where we
don't get a nullifier is almost trivially used.

I have kept `get_note_hash_and_optionally_a_nullifier` in nr intact as I
wasn't sure this desired changing. If so, I will do that in another PR
to keep things relatively isolated.

Resolves #11821

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>