fix: Switch to `noir-repo` context for cache content hashing (#12784)
There was a problem with the calculation of the content hash for
`noir/noir-repo`, because `cache_content_hash` uses `git ls-tree` to
list objects, and because `noir-repo` is a subdirectory it was seen as
effectively empty, short circuiting the builds. The PR changes this so
that the content hash is actually calculated in the `noir-repo`, where
things are visible:

* Add `REPO_ROOT` to `ci3/cache_content_hash` to specify the directory
where the git commands should run, so that we can evaluate paths in the
`noir-repo`
* Split `noir/.rebuild_patterns[_tests]` into `noir/.rebuild_patterns`
and `noir/.noir-repo.rebuild_patterns[_tests]`, with the latter having
their paths modified so they are rooted at `noir/noir-repo`.
* Change `noir/bootstrap.sh hash` to calculate multiple hashes, passing
the correct context with the different pattern files, and hashing the
final results together with `hash_str`.
* Add `noir/noir-repo.patch` to `noir/.rebuild_patterns`, but not
`noir/noir-repo-ref`, because that would be equivalent to treating any
change in the `noir-repo` as cache invalidation, almost defeating the
purpose of `.noir-repo.rebuild_patterns` (although having it would still
rebuild after new commits have been added to the `noir-repo`, it would
_also_ rebuild if nothing just the docs changed between two nightly
builds)
* Change all `bootstrap.sh` scripts where e.g.
`../../noir/.rebuild_patterns` was used directly to call
`noir/bootstrap.sh hash` instead.

### Original approach

The first commits in this PR took a different approach, trying to hide
the context switching in `cache_content_hash` itself, by changing it to:
* look for sub repositories in the `.rebuild_patterns`, e.g.
`^noir/noir-repo/.git`,
* loop through all such sub repositories, (starting from the top level
`.git` which is implicitly included in the loop)
* filter the rebuild patterns to the ones that start with the repository
path as a prefix
* calculate the hash of the files that match in that repository
* aggregate the hashes of across all the repos

The benefit of this approach in retrospect was that if e.g.
`avm-transpiler/bootstrap.sh` called `cache_content_hash
.rebuild_patterns ../../noir/.rebuild_patterns` then it would have kept
working, because `../../noir/.rebuild_patterns` had all the information
needed to pick the right context.

It was judged that `cache_content_hash` should not be burdened with this
kind of responsibility, so now we have to be vigilant and know when to
call `<project>/bootstrap.sh hash` and when we're okay to just use
`<project>/.rebuild_patterns` directly.