feat(avm): vm2 initial context (#12972)
This adds the initial work for `context` and `context_stack` to vm2. The
interfaces will still need to be updated in the future, but it sets up
the points where the context events will happen (i.e. at the end of the
main execution loop and within `call`).

This also outlines the contents of the two events albeit commented out
for now (until we have the supported inputs in the context itself)