feat: use magic address for fee juice portal in msgs (#13241)
When working with @alexghr on the masternet, we encountered pain points
around the initialization of the fee portal. Namely, that it required an
L1 tx to fund the bridge and then an L2 transactions to initialize it to
link them correctly.

This is a bit of a pain since it requires us to actually produces blocks
in order to allow transactions. Meaning that if we are using an initial
sequencer set, we run into a hurdle as they need to produce a block for
us first where we are making this link. Which means that it require some
level of activity.

Also, the L2 is not protected by any form of access control. So anyone
that is a little fast could just initialise it to whatever portal they
want and then mint themselves fees at will, without them actually being
backed.

To solve this hurdle, we suggest to let the inbox handle the fee portal
as a special case, and instead of the `msg.sender` use the magical
address that it also has inside the rollup. Since this would collide
with a pre-compile on ethereum sending the tx to the inbox, it seems
sane to assume that it will not happen. Also we assume that no-one have
the private key for these low value addresses (0x5).

This allows us to reduce the funding needed to 0. Since we do not have
to initialize the fee juice on L2 anymore 🎉. This is very neat as it
means that we don't need to have this first L2 transaction. Separately,
it showcased an issue as the current pre-funded accounts do not actually
bridge any assets. They essentially have completely unbacked fees, which
walls apart as soon as some of these are used to pay transactions fees
as they cannot be moved from the portal as the portal have nothing. This
issue have not manifested since the default value that was minted to the
fee juice for initialization was gigantic meaning that it was not used
completely by any tests so even if unbacked, they just ended up spending
someone elses collateral (the fee portals own collateral)

This means that when setting up a genesis state, we are now also
specifying the amount that we will need to be properly funded. As part
of the deployment, we will then mint this amount to the fee portal to
properly back those positions.