feat: dsl layer for recursive avm verifier v2 (#13362)
This PR creates the DSL layer for the goblinized avm recursive verifier
version 2.
The core logic lies in avm2_recursion_constraint.cpp and the goblinized
v2 verifier is activated in routine process_avm_recursion_constraints()
of acir_format.cpp.

As a consequence of enabling this verifier in the DSL layer, the AVM
integration tests are hooked with this version. The AVM integration test
was upgraded as it now integrates a clientIvc proof as well.
In addition, the new avm recursive verifier was added in the
rollup_ivc_integration test.

During this work, we hit an issue related to ipa claim/proof which is
not yet properly handled in the DSL layer which required us to
temporarily remove them from the goblinized avm recursive verifier. This
unfortunately led to the unit test AvmRecursiveTests.GoblinRecursion
being temporarily disabled.

---------

Co-authored-by: ledwards2225 <l.edwards.d@gmail.com>
Co-authored-by: sirasistant <sirasistant@gmail.com>